project/requirements.txt
openai==1.3.5
python-dotenv==1.0.0
streamlit==1.29.0
sounddevice==0.4.6
scipy==1.11.4
numpy==1.26.2
soundfile==0.12.1
pyaudio==0.2.14
wave==0.0.2
google-cloud-speech==2.21.0
google-cloud-storage==2.13.0